+++ config.json
@@ -11,7 +11,7 @@
   "NumberOfBuffers": 2,
   "CTBufferSize": 25,
   "StAndFwBufferSize": 1600,
-  "Routing": "CutThrough",
+  "Routing": "Store&Forward",
   "TracingEnabled": false,
   "RPCTimeoutMaxBackoff": 1000,
   "RPCTimeoutDuration": 60000,
