+++ sdk/go_sdk/destination.go
@@ -20,7 +20,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-package sdk
+package go_sdk
 
 import (
 	"context"
