+++ tests/go.mod
@@ -9,7 +9,7 @@ replace (
 	github.com/ease-lab/xdt/proto/downXDT => ../proto/downXDT
 	github.com/ease-lab/xdt/proto/upXDT => ../proto/upXDT
 	github.com/ease-lab/xdt/sQP => ../sQP
-	github.com/ease-lab/xdt/sdk => ../sdk
+	github.com/ease-lab/xdt/sdk/go_sdk => ../sdk/go_sdk
 	github.com/ease-lab/xdt/transport => ../transport
 	github.com/ease-lab/xdt/utils => ../utils
 )
@@ -20,7 +20,7 @@ require (
 	github.com/ease-lab/vhive/utils/tracing/go v0.0.0-20210624210547-e0cd5d053491
 	github.com/ease-lab/xdt/dQP v0.0.0-00010101000000-000000000000
 	github.com/ease-lab/xdt/sQP v0.0.0-00010101000000-000000000000
-	github.com/ease-lab/xdt/sdk v0.0.0-00010101000000-000000000000
+	github.com/ease-lab/xdt/sdk/go_sdk v0.0.0-00010101000000-000000000000
 	github.com/ease-lab/xdt/utils v0.0.0-00010101000000-000000000000
 	github.com/sirupsen/logrus v1.8.1
 	golang.org/x/net v0.0.0-20210415231046-e915ea6b2b7d
