+++ tests/integration_test.go
@@ -33,7 +33,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/ease-lab/xdt/sdk"
+	sdk "github.com/ease-lab/xdt/sdk/go_sdk"
 	"golang.org/x/net/http2"
 	"golang.org/x/net/http2/h2c"
 	pkgnet "knative.dev/pkg/network"
