+++ user-functions/gx/gx.go
@@ -26,11 +26,10 @@ import (
 	"flag"
 	"os"
 
-	"github.com/ease-lab/xdt/sdk"
-
 	ctrdlog "github.com/containerd/containerd/log"
 	log "github.com/sirupsen/logrus"
 
+	sdk "github.com/ease-lab/xdt/sdk/go_sdk"
 	"github.com/ease-lab/xdt/utils"
 )
